# API/server/nodejs-server-server-generated/api/openapi.yaml
openapi: 3.0.0
info:
  title: Machine fault diagnosis
  description: List of top level server APIs
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/zhuofusong/machine-fault-diagnosis/1.0.0
  description: Machine-fault-diagnosis backend server
security:
- application:
  - read
  - write
...
paths:
  /modelFlow:
    get:
      tags:
      - model flow
      summary: get a list of all model flow charts
      description: Get a list of all model flow charts
      operationId: modelFlowGET
      parameters:
      - name: modelFlowSearchString
        in: query
        description: the string for searching for model flow charts
        required: false
        allowEmptyValue: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: a list of model flow charts returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/modelFlowChartMeta'
                x-content-type: application/json
        "400":
          description: input is invalid
        "404":
          description: No model flow chart satisfies the search criteria
        "500":
          description: internal server error
      x-swagger-router-controller: ModelFlow
    delete:
      tags:
      - model flow
      summary: delete a list of model flow charts
      description: delete a list of model flow charts
      operationId: modelFlowDELETE
      parameters:
      - name: modelFlowSearchString
        in: query
        description: the string for searching for model flow charts
        required: false
        allowEmptyValue: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: the list of model flow charts deleted
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
        "400":
          description: input is invalid
        "404":
          description: No model flow chart satisfies the search criteria
        "500":
          description: internal server error
      x-swagger-router-controller: ModelFlow
...
  /modelFlow/node/{modelFlowId}/{nodeId}:
    get:
      tags:
      - model flow chart node
      summary: retrieve a node's information in a model flow chart
      description: retrieve a node's information in a model flow chart
      operationId: modelFlowNodeModelFlowIdNodeIdGET
      parameters:
      - name: modelFlowId
        in: path
        description: model flow chart id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: nodeId
        in: path
        description: node's id in a model flow chart
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: node returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/modelFlowChartNodeMeta'
                x-content-type: application/json
        "400":
          description: input is invalid
        "404":
          description: model flow chart node not found
        "500":
          description: internal server error
      x-swagger-router-controller: ModelFlowChartNode
    put:
      tags:
      - model flow chart node
      summary: update a node's information in a model flow chart
      description: update a node's information in a model flow chart
      operationId: modelFlowNodeModelFlowIdNodeIdPUT
      parameters:
      - name: modelFlowId
        in: path
        description: model flow chart id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: nodeId
        in: path
        description: node's id in a model flow chart
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: node updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/modelFlowChartNodeMeta'
                x-content-type: application/json
        "400":
          description: input is invalid
        "404":
          description: model flow chart not found. create one first.
        "500":
          description: internal server error
      x-swagger-router-controller: ModelFlowChartNode
    post:
      tags:
      - model flow chart node
      summary: create a node's information in a model flow chart
      description: create a node's information in a model flow chart
      operationId: modelFlowNodeModelFlowIdNodeIdPOST
      parameters:
      - name: modelFlowId
        in: path
        description: model flow chart id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: nodeId
        in: path
        description: node's id in a model flow chart
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: node created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/modelFlowChartNodeMeta'
                x-content-type: application/json
        "400":
          description: input is invalid
        "404":
          description: model flow chart not found. create one first.
        "500":
          description: internal server error
      x-swagger-router-controller: ModelFlowChartNode
    delete:
      tags:
      - model flow chart node
      summary: delete a node's information in a model flow chart
      description: delete a node's information in a model flow chart
      operationId: modelFlowNodeModelFlowIdNodeIdDELETE
      parameters:
      - name: modelFlowId
        in: path
        description: model flow chart id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: nodeId
        in: path
        description: node's id in a model flow chart
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: node deleted
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
        "400":
          description: input is invalid
        "404":
          description: model flow chart node not found
        "500":
          description: internal server error
      x-swagger-router-controller: ModelFlowChartNode
  /modelFlow/node/{modelFlowId}/{edgeId}:
    get:
      tags:
      - model flow chart edge
      summary: retrieve an edge's information in a model flow chart
      description: retrieve an edge's information in a model flow chart
      operationId: modelFlowNodeModelFlowIdEdgeIdGET
      parameters:
      - name: modelFlowId
        in: path
        description: model flow chart id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: edgeId
        in: path
        description: edge's id in a model flow chart
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: edge returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/modelFlowChartEdgeMeta'
                x-content-type: application/json
        "400":
          description: input is invalid
        "404":
          description: model flow chart edge not found
        "500":
          description: internal server error
      x-swagger-router-controller: ModelFlowChartEdge
    put:
      tags:
      - model flow chart edge
      summary: update an edge's information in a model flow chart
      description: update an edge's information in a model flow chart
      operationId: modelFlowNodeModelFlowIdEdgeIdPUT
      parameters:
      - name: modelFlowId
        in: path
        description: model flow chart id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: edgeId
        in: path
        description: edge's id in a model flow chart
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: edge updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/modelFlowChartEdgeMeta'
                x-content-type: application/json
        "400":
          description: input is invalid
        "404":
          description: model flow chart not found. create one first.
        "500":
          description: internal server error
      x-swagger-router-controller: ModelFlowChartEdge
    post:
      tags:
      - model flow chart edge
      summary: create an edge's information in a model flow chart
      description: create an edge's information in a model flow chart
      operationId: modelFlowNodeModelFlowIdEdgeIdPOST
      parameters:
      - name: modelFlowId
        in: path
        description: model flow chart id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: edgeId
        in: path
        description: edge's id in a model flow chart
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: edge created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/modelFlowChartEdgeMeta'
                x-content-type: application/json
        "400":
          description: input is invalid
        "404":
          description: model flow chart not found. create one first.
        "500":
          description: internal server error
      x-swagger-router-controller: ModelFlowChartEdge
    delete:
      tags:
      - model flow chart edge
      summary: delete an edge's information in a model flow chart
      description: delete an edge's information in a model flow chart
      operationId: modelFlowNodeModelFlowIdEdgeIdDELETE
      parameters:
      - name: modelFlowId
        in: path
        description: model flow chart id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: edgeId
        in: path
        description: edge's id in a model flow chart
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: edge deleted
          content:
            application/json:
              schema:
                type: object
                x-content-type: application/json
        "400":
          description: input is invalid
        "404":
          description: model flow chart edge not found
        "500":
          description: internal server error
      x-swagger-router-controller: ModelFlowChartEdge
...
components:
  schemas:
    modelFlowChartMeta:
      required:
      - id
      - kind
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        createdTime:
          type: number
        lastModifiedTime:
          type: number
        author:
          type: string
        editMode:
          type: string
        config:
          type: object
        comments:
          type: string
      example:
        lastModifiedTime: 6.02745618307040320615897144307382404804229736328125
        comments: comments
        kind: kind
        author: author
        editMode: editMode
        createdTime: 0.80082819046101150206595775671303272247314453125
        id: id
        config: {}
...
    modelFlowChartNodeMeta:
      required:
      - id
      - kind
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
          description: what kind of node it is. A 'simple' node or a supernode.
        createdTime:
          type: number
        modifiedTime:
          type: number
        author:
          type: string
        comments:
          type: string
      example:
        modifiedTime: 6.02745618307040320615897144307382404804229736328125
        comments: comments
        kind: kind
        author: author
        createdTime: 0.80082819046101150206595775671303272247314453125
        id: id
...
    modelFlowChartEdgeMeta:
      required:
      - id
      - kind
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
          description: the kind of string based on which
        createdTime:
          type: number
        modifiedTime:
          type: number
        author:
          type: string
        comments:
          type: string
      example:
        modifiedTime: 6.02745618307040320615897144307382404804229736328125
        comments: comments
        kind: kind
        author: author
        createdTime: 0.80082819046101150206595775671303272247314453125
        id: id
...
  securitySchemes:
    application:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://example.com/oauth/token
          scopes:
            write: allows modifying resources
            read: allows reading resources